Refactor code to separate file and integrate with latest changes
